Update and rename other requirements section to core requirements
