+++ apps/web/src/pages/Pools/EditPoolPage/components/CoreRequirementsSection/Display.tsx
